Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "black>=26.1.0",
    "fastapi>=0.128.2",
    "mistralai>=1.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
